develop/modules/antenna-autoassigner/requirements.txt
confluent-kafka[avro]
skyfield
numpy
psycopg2-binary
PyYAML